// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  ADMIN: 'ADMIN',
  VENUE_MANAGER: 'VENUE_MANAGER',
  DUTY_MANAGER: 'DUTY_MANAGER',
  BOUNCER: 'BOUNCER'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const BanStatus = {
  PENDING: 'PENDING',
  APPROVED: 'APPROVED',
  DENIED: 'DENIED',
  EXPIRED: 'EXPIRED'
} as const

export type BanStatus = (typeof BanStatus)[keyof typeof BanStatus]